Reuse testing code through TestDynamoDbService
