+++ tempest2-testing-internal/src/main/kotlin/app/cash/tempest2/testing/internal/TestDynamoDbService.kt
@@ -1,0 +1,66 @@
+package app.cash.tempest2.testing.internal
+
+import app.cash.tempest2.testing.TestDynamoDbClient
+import app.cash.tempest2.testing.TestDynamoDbServer
+import app.cash.tempest2.testing.TestTable
+import com.google.common.util.concurrent.AbstractIdleService
+import java.util.concurrent.ConcurrentHashMap
+
+/**
+ * This Guava service spins up a DynamoDB server in tests. It shares the server across tests and
+ * keeps the server running until the process exits.
+ */
+class TestDynamoDbService(
+  val client: TestDynamoDbClient,
+  val server: TestDynamoDbServer
+) : AbstractIdleService() {
+
+  override fun startUp() {
+    server.startIfNeeded()
+    client.startAsync()
+    client.awaitRunning()
+  }
+
+  override fun shutDown() {
+    client.stopAsync()
+    client.awaitTerminated()
+  }
+
+  private fun TestDynamoDbServer.startIfNeeded() {
+    if (runningServers.contains(id)) {
+      log.info { "$id already running, not starting anything" }
+      return
+    }
+    log.info { "starting $id" }
+    startAsync()
+    awaitRunning()
+    Runtime.getRuntime().addShutdownHook(
+      Thread {
+        log.info { "stopping $id" }
+        stopAsync()
+        awaitTerminated()
+      }
+    )
+    runningServers.add(id)
+  }
+
+  companion object {
+    private val defaultPorts = ConcurrentHashMap<String, Int>()
+    private fun defaultPort(key: String): Int {
+      // Only pick random port once to share one test server with multiple tests.
+      return defaultPorts.getOrPut(key, ::pickRandomPort)
+    }
+
+    private val runningServers = ConcurrentHashMap.newKeySet<String>()
+    private val log = getLogger<TestDynamoDbService>()
+
+    @JvmStatic
+    fun create(serverFactory: TestDynamoDbServer.Factory<*>, tables: List<TestTable>, port: Int?): TestDynamoDbService {
+      val port = port ?: defaultPort(serverFactory.toString())
+      return TestDynamoDbService(
+        DefaultTestDynamoDbClient(tables, port),
+        serverFactory.create(port)
+      )
+    }
+  }
+}
